+++ src/main/java/com/ontotext/refine/client/command/processes/GetProcessesCommand.java
@@ -11,13 +11,13 @@ import com.ontotext.refine.client.exceptions.RefineException;
 import com.ontotext.refine.client.util.HttpParser;
 import com.ontotext.refine.client.util.JsonParser;
 import java.io.IOException;
+import java.io.InputStream;
 import java.net.URI;
 import java.util.Collection;
 import org.apache.http.HttpResponse;
 import org.apache.http.HttpStatus;
 import org.apache.http.client.methods.HttpUriRequest;
 import org.apache.http.client.methods.RequestBuilder;
-
 
 /**
  * A command for retrieving the processes that are currently running for project.
@@ -27,7 +27,7 @@ import org.apache.http.client.methods.RequestBuilder;
 public class GetProcessesCommand implements RefineCommand<GetProcessesCommandResponse> {
 
   private static final TypeReference<Collection<ProjectProcess>> PROCESSES_TYPE =
-      new TypeReference<Collection<ProjectProcess>>() {};
+      new TypeReference<>() {};
 
   private final String project;
 
@@ -60,10 +60,12 @@ public class GetProcessesCommand implements RefineCommand<GetProcessesCommandRes
   @Override
   public GetProcessesCommandResponse handleResponse(HttpResponse response) throws IOException {
     HttpParser.HTTP_PARSER.assureStatusCode(response, HttpStatus.SC_OK);
-    JsonNode json = JsonParser.JSON_PARSER.parseJson(response.getEntity().getContent());
-    Collection<ProjectProcess> processes =
-        JsonParser.JSON_PARSER.read(json.findValue("processes").toString(), PROCESSES_TYPE);
-    return new GetProcessesCommandResponse(processes);
+    try (InputStream stream = response.getEntity().getContent()) {
+      JsonNode json = JsonParser.JSON_PARSER.parseJson(stream);
+      Collection<ProjectProcess> processes =
+          JsonParser.JSON_PARSER.read(json.findValue("processes").toString(), PROCESSES_TYPE);
+      return new GetProcessesCommandResponse(processes);
+    }
   }
 
   /**