+++ src/main/java/com/ontotext/refine/client/command/rdf/ExportRdfResponse.java
@@ -1,5 +1,7 @@
 package com.ontotext.refine.client.command.rdf;
 
+import java.io.File;
+import java.io.InputStream;
 
 /**
  * Holds the result from export RDF commands.
@@ -10,6 +12,8 @@ public class ExportRdfResponse {
 
   private String project;
   private String result;
+  private InputStream resultStream;
+  private File resultFile;
 
   public String getProject() {
     return project;
@@ -26,6 +30,24 @@ public class ExportRdfResponse {
 
   ExportRdfResponse setResult(String result) {
     this.result = result;
+    return this;
+  }
+
+  public InputStream getResultStream() {
+    return resultStream;
+  }
+
+  ExportRdfResponse setResultStream(InputStream resultStream) {
+    this.resultStream = resultStream;
+    return this;
+  }
+
+  public File getResultFile() {
+    return resultFile;
+  }
+
+  ExportRdfResponse setResultFile(File resultFile) {
+    this.resultFile = resultFile;
     return this;
   }
 }