+++ src/main/java/com/ontotext/refine/client/command/version/GetVersionCommand.java
@@ -2,6 +2,7 @@ package com.ontotext.refine.client.command.version;
 
 import static com.ontotext.refine.client.util.HttpParser.HTTP_PARSER;
 import static com.ontotext.refine.client.util.JsonParser.JSON_PARSER;
+import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.apache.http.HttpHeaders.ACCEPT;
 import static org.apache.http.HttpStatus.SC_OK;
 import static org.apache.http.entity.ContentType.APPLICATION_JSON;
@@ -15,7 +16,6 @@ import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.HttpUriRequest;
 import org.apache.http.client.methods.RequestBuilder;
 import org.apache.http.util.EntityUtils;
-
 
 /**
  * A command to retrieve Refine server version information.
@@ -53,7 +53,7 @@ public class GetVersionCommand implements RefineCommand<GetVersionResponse> {
   @Override
   public GetVersionResponse handleResponse(HttpResponse response) throws IOException {
     HTTP_PARSER.assureStatusCode(response, SC_OK);
-    String responseBody = EntityUtils.toString(response.getEntity());
+    String responseBody = EntityUtils.toString(response.getEntity(), UTF_8);
     return parseGetVersionResponse(responseBody);
   }